branch merged

+++ src/core/controller/User/userController.js
@@ -1,5 +1,11 @@
 import bcrypt from "bcryptjs";
-import { userModel, postModel, commentModel, likeModel } from "../../models";
+import {
+  userModel,
+  postModel,
+  commentModel,
+  likeModel,
+  relationshipModel,
+} from "../../models";
 import {
   onSuccess,
   onError,
@@ -61,7 +67,7 @@ const updateUser = async (request, response) => {
       request.body;
     const user = await userModel.findByPk(request.params.id);
     if (!user) {
-      return response.status(404).send("User not found");
+      return sendResponse(onError(404, "User not found"), response);
     }
     const salt = await bcrypt.genSalt(10);
     const hashedPassword = await bcrypt.hash(password, salt);
@@ -84,6 +90,12 @@ const updateUser = async (request, response) => {
 
 const deleteUser = async (request, response) => {
   try {
+    const userExists = await userModel.findOne({
+      where: { id: request.params.id },
+    });
+    if (!userExists) {
+      return sendResponse(onError(404, "User not found"), response);
+    }
     const deleteLikes = await likeModel.destroy({
       where: { userId: request.params.id },
     });
@@ -93,10 +105,32 @@ const deleteUser = async (request, response) => {
     const deletePosts = await postModel.destroy({
       where: { userId: request.params.id },
     });
+    const deleteRelationships = await relationshipModel.destroy({
+      where: {
+        followerId: request.params.id,
+        followedUserId: request.params.id,
+      },
+    });
     const deletedUser = await userModel.destroy({
       where: { id: request.params.id },
     });
     return sendResponse(onSuccess(200, "User deleted", deletedUser), response);
+  } catch (error) {
+    globalCatch(request, error);
+    return sendResponse(
+      onError(500, messageResponse.ERROR_FETCHING_DATA),
+      response
+    );
+  }
+};
+
+const getUser = async (request, response) => {
+  try {
+    const user = await userModel.findByPk(request.params.id);
+    if (!user) {
+      return sendResponse(onError(404, "User not found"), response);
+    }
+    return sendResponse(onSuccess(200, "User details", user), response);
   } catch (error) {
     globalCatch(request, error);
     return sendResponse(
@@ -119,4 +153,4 @@ const getUsers = async (request, response) => {
   }
 };
 
-export default { createUser, deleteUser, updateUser, getUsers };
+export default { createUser, deleteUser, updateUser, getUsers, getUser };
